+++ sdk/metric/manual_reader.go
@@ -50,7 +50,7 @@ func NewManualReader(opts ...ManualReaderOption) *ManualReader {
 		temporalitySelector: cfg.temporalitySelector,
 		aggregationSelector: cfg.aggregationSelector,
 	}
-	r.externalProducers.Store([]Producer{})
+	r.externalProducers.Store(cfg.producers)
 	return r
 }
 
@@ -64,23 +64,6 @@ func (mr *ManualReader) register(p sdkProducer) {
 	}
 }
 
-// RegisterProducer stores the external Producer which enables the caller
-// to read metrics on demand.
-//
-// This method is safe to call concurrently.
-func (mr *ManualReader) RegisterProducer(p Producer) {
-	mr.mu.Lock()
-	defer mr.mu.Unlock()
-	if mr.isShutdown {
-		return
-	}
-	currentProducers := mr.externalProducers.Load().([]Producer)
-	newProducers := []Producer{}
-	newProducers = append(newProducers, currentProducers...)
-	newProducers = append(newProducers, p)
-	mr.externalProducers.Store(newProducers)
-}
-
 // temporality reports the Temporality for the instrument kind provided.
 func (mr *ManualReader) temporality(kind InstrumentKind) metricdata.Temporality {
 	return mr.temporalitySelector(kind)
@@ -89,13 +72,6 @@ func (mr *ManualReader) temporality(kind InstrumentKind) metricdata.Temporality 
 // aggregation returns what Aggregation to use for kind.
 func (mr *ManualReader) aggregation(kind InstrumentKind) aggregation.Aggregation { // nolint:revive  // import-shadow for method scoped by type.
 	return mr.aggregationSelector(kind)
-}
-
-// ForceFlush is a no-op, it always returns nil.
-//
-// This method is safe to call concurrently.
-func (mr *ManualReader) ForceFlush(context.Context) error {
-	return nil
 }
 
 // Shutdown closes any connections and frees any resources used by the reader.
@@ -183,6 +159,7 @@ func (r *ManualReader) MarshalLog() interface{} {
 type manualReaderConfig struct {
 	temporalitySelector TemporalitySelector
 	aggregationSelector AggregationSelector
+	producers           []Producer
 }
 
 // newManualReaderConfig returns a manualReaderConfig configured with options.